django_internal_project/requirements.txt
asgiref==3.8.1
Cerberus==1.3.5
Django==5.0.8
djangorestframework==3.15.2
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
django-environ==0.11.2


environ~=1.0